requirements.txt
Flask
deepface
retinaface
pillow
numpy
opencv-python-headless
tensorflow
keras
